# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Lead:
    properties:
      assigned_to:
        type: integer
      category_id:
        type: integer
      contact:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      status:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Staff:
    properties:
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      position:
        type: string
      profile:
        description: JSONB field for custom attributes
        type: string
      role:
        $ref: '#/definitions/models.StaffRole'
      tenant_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.StaffRole:
    enum:
    - admin
    - manager
    - employee
    type: string
    x-enum-varnames:
    - RoleStaffAdmin
    - RoleStaffManager
    - RoleStaffEmployee
...
host: localhost:3000
info:
  contact:
    email: support@kontena.io
    name: API Support
  description: Multi-tenant CRM API built with Golang and Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kontena CRM API
  version: "1.0"
paths:
...
  /leads:
    get:
      consumes:
      - application/json
      description: Get all leads for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Lead'
            type: array
      summary: Get all leads
      tags:
      - leads
    post:
      consumes:
      - application/json
      description: Create a new lead for the current tenant
      parameters:
      - description: Lead information
        in: body
        name: lead
        required: true
        schema:
          $ref: '#/definitions/models.Lead'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Lead'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a lead
      tags:
      - leads
  /leads/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a lead by ID for the current tenant
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a lead
      tags:
      - leads
    get:
      consumes:
      - application/json
      description: Get a lead by ID for the current tenant
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lead'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a lead
      tags:
      - leads
    patch:
      consumes:
      - application/json
      description: Update an existing lead for the current tenant
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lead information
        in: body
        name: lead
        required: true
        schema:
          $ref: '#/definitions/models.Lead'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lead'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a lead
      tags:
      - leads
  /staff:
    get:
      consumes:
      - application/json
      description: Get all staff members for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Staff'
            type: array
      summary: Get all staff members
      tags:
      - staff
    post:
      consumes:
      - application/json
      description: Create a new staff member for the current tenant
      parameters:
      - description: Staff member information
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.Staff'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a staff member
      tags:
      - staff
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a staff member by ID for the current tenant
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a staff member
      tags:
      - staff
    get:
      consumes:
      - application/json
      description: Get a staff member by ID for the current tenant
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a staff member
      tags:
      - staff
    patch:
      consumes:
      - application/json
      description: Update an existing staff member for the current tenant
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      - description: Staff member information
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.Staff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a staff member
      tags:
      - staff
...
swagger: "2.0"